refactor(ui): bryt ut pure logik ur DataTable + firma-settings-panel (#62)
Behåller publik API via re-exports → inga importörer eller tester ändras.

- data-table.tsx (814→647 rad): pure typer + sort/filter/group/prefs-funktioner
  flyttade till data-table-logic.ts (187 rad). Komponenten + dess ~25 sub-
  komponenter blir kvar; logiken (den tungt testade, ramverks-agnostiska
  kärnan) får ett eget hem (SRP).
- firma-settings-panel.tsx (652→594 rad): GitHub-token-/repo-validering +
  proxy-tester flyttade till firma-settings-net.ts (72 rad). Panelen var redan
  uppdelad i sub-komponenter; nätverkslogiken separeras nu från React.

OBS: filerna var redan lint-rena (max-lines-per-function gäller per FUNKTION,
inte per fil) — de fanns INTE i eslint-suppressions.json, så det finns inget
att pruna. Värdet är modul-kohesion/SRP, inte lint-skuld. Beteende oförändrat
(2262 tester, data-table/firma-settings-panel-testerna gröna via re-exports).
